refactor: move CooldownTimer component to a dedicated component file
