+++ src/components/MessageInput/MessageInputFlat.tsx
@@ -8,7 +8,6 @@ import zipObject from 'lodash/zipObject';
 import { nanoid } from 'nanoid';
 
 import { EmojiPicker } from './EmojiPicker';
-import { CooldownTimer as DefaultCooldownTimer } from './hooks/useCooldownTimer';
 import {
   EmojiIconLarge as DefaultEmojiIcon,
   EmojiPickerIcon as DefaultEmojiPickerIcon,
@@ -34,6 +33,7 @@ import { useMessageInputContext } from '../../context/MessageInputContext';
 import { useComponentContext } from '../../context/ComponentContext';
 
 import type { DefaultStreamChatGenerics } from '../../types/types';
+import { CooldownTimer as DefaultCooldownTimer } from './CooldownTimer';
 
 export const MessageInputFlat = <
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics
