+++ src/components/MessageInput/index.ts
@@ -8,3 +8,4 @@ export * from './MessageInputFlat';
 export * from './MessageInputSmall';
 export * from './QuotedMessagePreview';
 export * from './UploadsPreview';
+export * from './CooldownTimer';
